Add one or two important points in part1
(And change 1.6.2 to 1.6.3 because, hey, I can. :)

The example shows CREATE EXTENSION failing, and then the installation
succeeding when the right libjvm_location is given. But that way, it is
not installed as an extension. Add the missing step to fix that.

Introduce the idea of a deployment descriptor file, just creating it
by hand for the two functions in part1. How to have the Java compiler
create it can be left for part2.