Change return types to VARCHAR in part1
This blog has actually found a PL/Java bug! The Postgres type CHAR
with no (n) means the same as CHAR(1). If PL/Java were checking that,
it would be returning the result truncated to 1 character. I will
have to open a PL/Java issue about that.

But of course that's not what's wanted here, so the return types
should be declared VARCHAR.