Use PL/Java annotations in blog part 2
(And change 1.6.2 to 1.6.3 because, hey, I can. :)

Demonstrate the use of PL/Java's @Function and @Trigger annotations
in the Java source to completely avoid writing SQL by hand to create
the functions and triggers.

It doesn't just save some work. It makes sure the SQL stays up to date
with the code. It tells you at compile time if the constructs don't
conform to Postgres rules, so you don't only find out later when
loading things into the database. If you write your Java in an IDE and
you have the PL/Java API jar on its path, the IDE can step you through
the applicable options while you code.