Add UPDATE to trigger events, fix table name
I think the trigger events must have been a typo in the blog.
The example noop update in the blog doesn't really encrypt anything
if the trigger doesn't happen on updates.

The processQuery method had the right table name as presented
in the blog, but not in the actual source file.